+++ lib/dart_periphery.dart
@@ -3,7 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 export 'src/dummy.dart' show DummyDev;
+export 'src/isolate_api.dart' show IsolateAPI;
 export 'src/errno.dart' show ERRNO, Errno, ErrnoNotFound;
+export 'src/hardware/utils/byte_buffer.dart' show checkCRC, crc8;
 export 'src/gpio.dart'
     show
         GPIOpolling,
@@ -82,6 +84,9 @@ export 'src/hardware/sht31.dart'
         sht31DefaultI2Caddress;
 export 'src/hardware/mlx90615.dart'
     show MLX90615, MLX90615result, MLX90615exception, mlx90615DefaultI2Caddress;
+export 'src/hardware/ds1307.dart'
+    show DS1307, DS1307exception, ds1307DefaultI2Caddress;
+export 'src/hardware/linux_local_time.dart' show setLinuxLocalTime;
 export 'src/hardware/mcp9808.dart'
     show MCP9808, MCP9808result, MCP9808exception, mcp9808DefaultI2Caddress;
 export 'src/hardware/sdc30.dart'
@@ -102,6 +107,9 @@ export 'src/hardware/tsl2591.dart'
         IntegrationTime,
         RawLuminosity,
         tsl2591DefaultI2Caddress;
+export 'src/hardware/pcf8591.dart'
+    show PFC8591, PFC8591exception, Pin, pcf8591DefaultI2Caddress;
+
 export 'src/hardware/utils/byte_buffer.dart' show BitOrder;
 export 'src/hardware/utils/uint.dart';
 export 'src/i2c.dart'